Integrated browserify and brought back module exports and node-style requires to promote modularity
