+++ index.html
@@ -17,17 +17,10 @@
     <!-- CSS -->
     <link href="static/css/style.css" rel="stylesheet" type="text/css" />
 
-    <!-- Axios -->
-    <script src="https://unpkg.com/axios/dist/axios.min.js"></script>
-
     <!-- Local scripts -->
-    <script src="script/client.js"></script>
-    <script src="script/model.js"></script>
-   <!--<script src="script/api.js"></script>-->
-    <script src="script/controller.js"></script>
-    <script src="script/view.js"></script>
+    <script src="script/bundle.js"></script>
 </head>
-<body onload="onPageLoad()">
+<body onload="main.onPageLoad()">
 <header>
     <div class="collapse bg-dark" id="navbarHeader">
         <div class="container">
