+++ package.json
@@ -5,7 +5,6 @@
   "dependencies": {
     "axios": "^0.18.0",
     "jest": "^24.8.0",
-    "jquery": "^3.4.1",
-    "requirejs": "^2.3.6"
+    "jquery": "^3.4.1"
   }
 }
